feat(hero,castle): Hero election/leadership system + new (Hero/faction-owned) castle system
Extracts the Hero system (election cycles, leadership points, Mobilization
Order, Dominion Points, capital-city Statues) and the new Hero/faction-owned
castle system (Dominion claim/tax, player-founded Nations, Siege raid teams)
onto a clean upstream base, isolated from the old guild-owned Exeloch/Sungold
castle system and from unrelated work mixed into the same dev tree.

See HERO_AND_CASTLE_SYSTEM_CHANGELOG.md for the full compact.sqlite3/aaemu_game
database changelog and a list of what was deliberately left out.

GuildDominionManager/IGuildDominionManager are included only as a compile-time
dependency (HousingManager and AdvanceGuardTowerStep branch on both castle
systems at the same call site) - not new guild-system feature work.

Build- and test-verified: AAEmu.Game, AAEmu.World and AAEmu.UnitTests all
compile clean; full suite passes 1997/1997.

+++ AAEmu.Game/Core/Packets/G2C/SCFactionMobilizationOrderPacket.cs
@@ -4,13 +4,23 @@ using AAEmu.Game.Core.Network.Game;
 namespace AAEmu.Game.Core.Packets.G2C;
 
 /// <summary>
-/// TODO: nothing constructs this packet yet.
+/// Faction-wide broadcast: a Hero has issued a Mobilization Order. Answers
+/// CSFactionIssuanceOfMobilizationOrderPacket / CSFactionMobilizationOrderPacket - see
+/// HeroManager.IssueMobilizationOrder.
 /// </summary>
 /// <remarks>
-/// Field order, widths and names come from the 10.0.2.13 client's serializer, which passes each
-/// value's name alongside the value:
+/// 2026-08-31: RTTI-decoded the real Unpack (FUN_39c60be0) - 3 fields at entry+0x10 (slot 0x88),
+/// entry+0x18 (slot 0x98, this project's established 8-byte-plain-value slot), entry+0x20 (fixed
+/// 128-byte string, the "name" field, XlStringSize-driven like other name fields in this codebase).
+/// The 0x10-to-0x18 offset gap is 8 bytes, so <c>type</c> was widened from the earlier guessed
+/// <c>short</c> to <c>ulong</c> to match - this is offset-gap evidence, not a decompiled primitive
+/// width, so treat it as best-effort (SC packet, cosmetic risk only per this project's risk model:
+/// worst case is a garbled broadcast, not a crash). <c>type2</c>'s width (slot 0x98) IS
+/// primitive-confirmed. Semantics: <c>type</c>/<c>type2</c> are unnamed in the decompile; this
+/// implementation uses type2 for the acting Hero's persistent character id and name for their
+/// display name, type left 0 (unconfirmed purpose).
 /// </remarks>
-public class SCFactionMobilizationOrderPacket(short @type, ulong @type2, string name) : GamePacket(SCOffsets.SCFactionMobilizationOrderPacket, 1)
+public class SCFactionMobilizationOrderPacket(ulong @type, ulong @type2, string name) : GamePacket(SCOffsets.SCFactionMobilizationOrderPacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
     {
